fix : schema 정리

+++ DungeonAPI/DungeonAPI/ModelDB/Item.cs
@@ -7,9 +7,9 @@ public class Item
 	public Int32 ItemId { get; set; }
 	public Int32 ItemCode { get; set; }
 	public Int32 ItemCount { get; set; }
-	public Int64 Attack { get; set; }
-	public Int64 Defence { get; set; }
-	public Int64 Magic { get; set; }
+	public Int32 Attack { get; set; }
+	public Int32 Defence { get; set; }
+	public Int32 Magic { get; set; }
 	public Byte EnhanceLevel { get; set; } = 0;
 	public Byte EnhanceTryCount { get; set; } = 0;
 	public Boolean IsDestructed { get; set; } = false;
